fix CI

+++ src/SimpleVotingSystem.sol
@@ -5,7 +5,6 @@ import {AccessControl} from "openzeppelin-contracts/contracts/access/AccessContr
 import {VotingNFT} from "./VotingNFT.sol";
 
 contract SimpleVotingSystem is AccessControl {
-
     bytes32 public constant FOUNDER_ROLE = keccak256("FOUNDER_ROLE");
     bytes32 public constant WITHDRAWER_ROLE = keccak256("WITHDRAWER_ROLE");
 
@@ -22,19 +21,19 @@ contract SimpleVotingSystem is AccessControl {
     VotingNFT public votingNFT;
 
     struct Candidate {
-        uint id;
+        uint256 id;
         string name;
         address payable candidateAddress;
-        uint voteCount;
+        uint256 voteCount;
     }
 
-    mapping(uint => Candidate) public candidates;
+    mapping(uint256 => Candidate) public candidates;
     mapping(address => bool) public voters;
-    uint[] private candidateIds;
+    uint256[] private candidateIds;
 
     event WorkflowStatusChange(WorkflowStatus previousStatus, WorkflowStatus newStatus);
-    event CandidateFunded(uint indexed candidateId, uint amount);
-    event FundsWithdrawn(address indexed withdrawer, uint amount);
+    event CandidateFunded(uint256 indexed candidateId, uint256 amount);
+    event FundsWithdrawn(address indexed withdrawer, uint256 amount);
 
     constructor() {
         _grantRole(DEFAULT_ADMIN_ROLE, msg.sender);
@@ -46,7 +45,7 @@ contract SimpleVotingSystem is AccessControl {
     function setWorkflowStatus(WorkflowStatus _status) public onlyRole(DEFAULT_ADMIN_ROLE) {
         WorkflowStatus previousStatus = workflowStatus;
         workflowStatus = _status;
-        
+
         if (_status == WorkflowStatus.VOTE) {
             voteStartTime = block.timestamp;
         }
@@ -59,31 +58,31 @@ contract SimpleVotingSystem is AccessControl {
         require(bytes(_name).length > 0, "Candidate name cannot be empty");
         require(_candidateAddress != address(0), "Candidate address cannot be zero");
 
-        uint candidateId = candidateIds.length + 1;
+        uint256 candidateId = candidateIds.length + 1;
         candidates[candidateId] = Candidate(candidateId, _name, payable(_candidateAddress), 0);
         candidateIds.push(candidateId);
     }
 
-    function fundCandidate(uint _candidateId) public payable onlyRole(FOUNDER_ROLE) {
+    function fundCandidate(uint256 _candidateId) public payable onlyRole(FOUNDER_ROLE) {
         require(workflowStatus == WorkflowStatus.FOUND_CANDIDATES, "Funding candidates is not open");
         require(_candidateId > 0 && _candidateId <= candidateIds.length, "Invalid candidate ID");
         require(msg.value > 0, "You must send some ether");
 
         Candidate storage candidate = candidates[_candidateId];
-        
-        (bool success, ) = candidate.candidateAddress.call{value: msg.value}("");
+
+        (bool success,) = candidate.candidateAddress.call{value: msg.value}("");
         require(success, "Transfer failed");
 
         emit CandidateFunded(_candidateId, msg.value);
     }
 
-    function vote(uint _candidateId) public {
+    function vote(uint256 _candidateId) public {
         require(workflowStatus == WorkflowStatus.VOTE, "Voting session is not open");
         require(block.timestamp >= voteStartTime + 1 hours, "Voting starts 1 hour after session open");
-        
+
         require(votingNFT.balanceOf(msg.sender) == 0, "You already have the voting NFT");
         require(!voters[msg.sender], "You have already voted");
-        
+
         require(_candidateId > 0 && _candidateId <= candidateIds.length, "Invalid candidate ID");
 
         voters[msg.sender] = true;
@@ -96,8 +95,8 @@ contract SimpleVotingSystem is AccessControl {
         require(workflowStatus == WorkflowStatus.COMPLETED, "Voting session is not completed");
         require(address(this).balance > 0, "No funds to withdraw");
 
-        uint amount = address(this).balance;
-        (bool success, ) = msg.sender.call{value: amount}("");
+        uint256 amount = address(this).balance;
+        (bool success,) = msg.sender.call{value: amount}("");
         require(success, "Withdraw transfer failed");
 
         emit FundsWithdrawn(msg.sender, amount);
@@ -107,11 +106,11 @@ contract SimpleVotingSystem is AccessControl {
         require(workflowStatus == WorkflowStatus.COMPLETED, "Voting session is not completed");
         require(candidateIds.length > 0, "No candidates available");
 
-        uint winningCandidateId = candidateIds[0];
-        uint maxVotes = candidates[winningCandidateId].voteCount;
+        uint256 winningCandidateId = candidateIds[0];
+        uint256 maxVotes = candidates[winningCandidateId].voteCount;
 
-        for (uint i = 1; i < candidateIds.length; i++) {
-            uint currentId = candidateIds[i];
+        for (uint256 i = 1; i < candidateIds.length; i++) {
+            uint256 currentId = candidateIds[i];
             if (candidates[currentId].voteCount > maxVotes) {
                 maxVotes = candidates[currentId].voteCount;
                 winningCandidateId = currentId;
@@ -121,16 +120,16 @@ contract SimpleVotingSystem is AccessControl {
         return candidates[winningCandidateId];
     }
 
-    function getTotalVotes(uint _candidateId) public view returns (uint) {
+    function getTotalVotes(uint256 _candidateId) public view returns (uint256) {
         require(_candidateId > 0 && _candidateId <= candidateIds.length, "Invalid candidate ID");
         return candidates[_candidateId].voteCount;
     }
 
-    function getCandidatesCount() public view returns (uint) {
+    function getCandidatesCount() public view returns (uint256) {
         return candidateIds.length;
     }
 
-    function getCandidate(uint _candidateId) public view returns (Candidate memory) {
+    function getCandidate(uint256 _candidateId) public view returns (Candidate memory) {
         require(_candidateId > 0 && _candidateId <= candidateIds.length, "Invalid candidate ID");
         return candidates[_candidateId];
     }
